update roles permission crud

+++ app/Http/Controllers/RoleController.php
@@ -10,21 +10,6 @@ use Yajra\DataTables\Facades\DataTables;
 
 class RoleController extends Controller
 {
-    // public function index(Request $request)
-    // {
-    //     if ($request->ajax()) {
-    //         $roles = Role::with('permissions')->get();
-    //         return response()->json(['roles' => $roles]);
-    //     }
-
-    //     $permissions = Permission::all();
-    //     $groupedPermissions = [];
-    //     foreach ($permissions as $perm) {
-    //         [$module, $action] = explode('.', $perm->name);
-    //         $groupedPermissions[$module][$action] = $perm;
-    //     }
-    //     return view('roles.index', compact('permissions', 'groupedPermissions'));
-    // }
 
     public function index(Request $request)
     {
@@ -47,13 +32,13 @@ class RoleController extends Controller
                     $editBtn = '';
                     $deleteBtn = '';
 
-                    // if (auth()->user()->can('roles.update')) {
+                    if (auth()->user()->can('roles.update')) {
                         $editBtn = "<button class='btn btn-sm btn-warning btn-edit' data-id='{$role->id}'><i class='ri-pencil-line'></i></button>";
-                    // }
+                    }
 
-                    // if (auth()->user()->can('roles.delete')) {
+                    if (auth()->user()->can('roles.delete')) {
                         $deleteBtn = "<button class='btn btn-sm btn-danger btn-delete' data-id='{$role->id}'><i class='ri-delete-bin-6-line'></i></button>";
-                    // }
+                    }
 
                     return $editBtn . ' ' . $deleteBtn;
                 })
@@ -83,9 +68,9 @@ class RoleController extends Controller
 
     public function show($id)
     {
-        // if (!Auth::user()->can('roles.read')) {
-        //     abort(403, 'Anda tidak punya akses untuk menyimpan roles');
-        // }
+        if (!Auth::user()->can('roles.read')) {
+            abort(403, 'Anda tidak punya akses untuk menyimpan roles');
+        }
         $role = Role::findOrFail($id);
         $permissions = $role->permissions->pluck('name');
 
@@ -98,9 +83,9 @@ class RoleController extends Controller
 
     public function update(Request $request, Role $role)
     {
-        // if (!Auth::user()->can('users.update')) {
-        //     abort(403, 'Anda tidak punya akses untuk mengedit user');
-        // }
+        if (!Auth::user()->can('users.update')) {
+            abort(403, 'Anda tidak punya akses untuk mengedit user');
+        }
         $role->update(['name' => $request->name]);
         $role->syncPermissions($request->permissions ?? []);
         return response()->json(['message' => 'Role updated']);
@@ -108,9 +93,9 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
-        // if (!Auth::user()->can('roles.delete')) {
-        //     abort(403, 'Anda tidak punya akses untuk mengedit user');
-        // }
+        if (!Auth::user()->can('roles.delete')) {
+            abort(403, 'Anda tidak punya akses untuk mengedit user');
+        }
         $role->delete();
         return response()->json(['message' => 'Role deleted']);
     }
